requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2
Pillow==10.1.0
